test(control-plane): make the services package compile under GOOS=windows
dev_service.go is build-tagged !windows (Windows gets a stub service),
but two pieces of its test surface were not:

- mockFileSystemAdapter lived in dev_service_test.go (!windows) while
  untagged test files (package_service_test.go, coverage_additional_
  test.go) use it. Moved to a new untagged mocks_fs_test.go.
- Three tests in the untagged coverage_additional_test.go exercised
  Unix-only DefaultDevService methods (loadDevEnvFile, startDevProcess,
  port helpers) that the Windows stub does not define. Moved into the
  !windows-tagged dev_service_test.go.

Before this, go vet / go test of internal/core/services failed to
compile on Windows. Also carries the one-line stub update from the
interpreter-probe change (the fake python in coverage_additional_test.go
now answers the -c version query); no other test logic changed.
